Enable relative imports in python tests
