+++ tests/python_tests/__init__.py
@@ -1,0 +1,3 @@
+"""Tests for the python interface, intended for `pytest`."""
+# Copyright (C) TeNPy Developers, Apache license
+pass
