+++ tests/python_tests/networks/__init__.py
@@ -1,0 +1,3 @@
+"""Tests for the `cyten.networks` subpackage"""
+# Copyright (C) TeNPy Developers, Apache license
+pass
