+++ tests/python_tests/networks/test_site.py
@@ -12,7 +12,7 @@ import cyten
 from cyten import backends
 from cyten.networks import site
 
-from conftest import random_symmetry_sectors
+from ..conftest import random_symmetry_sectors
 
 
 pytest.skip("site not yet revised", allow_module_level=True)  # TODO
